05-data-platforms/zoomcamp/pipeline/assets/ingestion/requirements.txt
pandas
pyarrow
python-dateutil
tzdata